PR review feedback - added --list, DATA_DIR should now be properly inferred if it's not provided, and moved some code back into benchmark_runner that is cli only.
